# Improve event normalization performance