Enhancement

## What does this PR do?

Improve event normalization performance

## Why is it important?

generalize/normalize is widely used in beats

## Checklist

- [x] My code follows the style guidelines of this project
- [x] ~~I have commented my code, particularly in hard-to-understand areas~~
- [x] ~~I have made corresponding changes to the documentation~~
- [x] ~~I have made corresponding change to the default configuration files~~
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.

## Author's Checklist

- [ ]

## How to test this PR locally

Benchmark 

## Related issues

- 

## Use cases


## Screenshots



## Logs

```
$ benchstat old.txt new.txt                                                                              
name             old time/op    new time/op    delta
Normalization-8     984ns ±16%     559ns ±13%  -43.16%  (p=0.000 n=20+19)

name             old alloc/op   new alloc/op   delta
Normalization-8    1.18kB ± 0%    0.74kB ± 0%  -37.41%  (p=0.000 n=20+20)

name             old allocs/op  new allocs/op  delta
Normalization-8      13.0 ± 0%       5.0 ± 0%  -61.54%  (p=0.000 n=20+20)
```