@urso 2020-12-09
Merged. The change will be available in 7.11. I will take care of backporting.

Thank you for your contribution.